mvebu: RB5009: update SERDES lane config
By dumping the registers from the Jailbroken ROSv7 it was discovered that
the SERDES lane config is:
0 = PCIE
1 = PCIE
2 = SERDES (XFI)
3 = USB3
4 = PCIE
5 = PCIE

PCIe lanes are left unconfigured as its not clear how the PCIe lanes are
divided as there are 2 PCIe controllers with different number of lanes.
So since RB5009 doesnt actually have anything connected to PCIe (I assume
they will launch a WiFi model that will have PCIe based card) just define
it as unconnected.

This enables the USB port to work in USB3.0 mode (5Gbps), wired networking
looks to need further config of the switch.

Signed-off-by: Robert Marko <robimarko@gmail.com>