common: miiphyutil: initialize mii_devs list if its not initialized
Currently, miiphy_get_dev_by_name() uses mii_devs list without checking
if its initialized at all.
This can cause various NULL pointer crashes if one tries for example to
get or use a DM MDIO device before ethernet is initialized.

This is because for example board late init is called before ethernet
init in the whole init process, and miiphy_init() which initializes the
mii_devs is part of the ethernet init.

This effectively prevents having MDIO subnode based devices like switches
as calling dm_scan_fdt_dev() would cause a NULL pointer crash.

So, to prevent this check if mii_devs is empty and if so initialize it.

Signed-off-by: Robert Marko <robimarko@gmail.com>