[wip] mvebu: RB5009: attach PHY-s to ports
Since its not crashing anymore during port probe if PHY is present then
attach the PHY-s to switch ports.

Add the missing port 1(2.5G one) but we must also add a disabled dummy
port as DSA uclass checks if reg number equals the loop position and
will just exit if not so.

Signed-off-by: Robert Marko <robimarko@gmail.com>